adds query functions
