+++ lib/store.js
@@ -1,75 +1,8 @@
 "use strict";
 
 var StoreStateBasePrototype = require("./store/proto");
-
-/**
- * Encapsulate state in a "read-only" object.
- */
-function createStoreState(props, observer, strict) {
-    var state = {};
-
-    var keeper = Object.keys(props).reduce(function(stateKeeper, key) {
-        var def = props[key];
-
-        Object.defineProperty(stateKeeper, key, {
-            enumerable: true,
-            get: function() {
-                return state[key];
-            },
-
-            set: function(val) {
-                if (def.type) {
-                    var err = def.type(val);
-
-                    if (err instanceof Error) {
-                        console.error("Validation failed for attribute %s with value %s", key, val);
-                    }
-
-                    if (strict && err instanceof Error) {
-                        console.error("Strict warning: not setting %s", key);
-                        return;
-                    }
-
-                }
-
-                var prevValue = state[key];
-                state[key] = val;
-                observer(key, val, prevValue);
-            }
-        });
-
-        return stateKeeper;
-    }, {});
-
-    return Object.freeze(keeper);
-}
-
-function getStateProps(ctor, props, key) {
-    return {
-        store: ctor,
-        name: key,
-        type: props[key].type,
-    };
-}
-
-function createStateProps(ctor, props) {
-    return Object.keys(props).reduce(function(StateProps, key) {
-
-        Object.defineProperty(StateProps, key, {
-            enumerable: true,
-            get: function() {
-                var properties = getStateProps(ctor, props, key);
-                properties.query = function(query) {
-                    properties._query = query;
-                    return properties;
-                };
-                return properties;
-            }
-        });
-
-        return StateProps;
-    }, {});
-}
+var createStateProps = require("./store/state-props");
+var createStoreState = require("./store/state");
 
 function createStore(storeName, props) {
     var ctor = function Store(dispatcher) {
