+++ lib/store/proto.js
@@ -1,6 +1,14 @@
 /**
  * Shared prototype of all Stores, e.g. baseclass.
  */
+ 
+function jsonPathHandler(store, path){
+    return function(key, value, orig){
+        return store.state.query(key, path);
+    };
+}
+ 
+ 
 var StoreStateBasePrototype = {
     getState: function getState() {
         return this.state;
@@ -10,6 +18,15 @@ var StoreStateBasePrototype = {
         for ( var key in this.state ){
             this.state[key] = null;
         }
+    },
+    
+    query: function(attr, name, handler){
+        if(typeof handler === "string")
+            handler = jsonPathHandler(this, handler);
+
+        this.observe(attr, function(key, value, orig){
+            this.__notify(name, handler(key, value, orig));            
+        }.bind(this));
     },
 
     observe: function(attr, handler) {
